requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-jazzmin==2.6.0
requests==2.30.0
